refactor config_window
- made config_window a module
  - mod.rs is config_window
  - sidebar and bottombar got dedicated modules and structs
