+++ src/ui/config_window/mod.rs
@@ -1,56 +1,37 @@
+pub mod bottombar;
+pub mod sidebar;
+
 use crate::dataloading::dataprovider::song_data_provider::{
-    SongChange, SongDataEdit, SongDataProvider, SongDataSource,
+    SongChange, SongDataEdit, SongDataSource,
 };
-use crate::traktor_api::{TRAKTOR_SERVER_DEFAULT_ADDR, TraktorNextMode, TraktorSyncMode};
-use crate::ui::widget::canvas_toggle::CanvasToggle;
+use crate::ui::config_window::bottombar::BottomBar;
+use crate::ui::config_window::sidebar::Sidebar;
 use crate::ui::widget::dynamic_text_input::DynamicTextInput;
-use crate::ui::widget::suggestion_text_input::SuggestionTextInput;
-use crate::ui::widget::{power_button, suggestion_text_input};
 use crate::ui::{material_icon, material_icon_sized};
 use crate::{DanceInterpreter, Message, Window};
 use iced::advanced::Widget;
 use iced::alignment::Vertical;
 use iced::border::Radius;
-use iced::widget::scrollable::{Direction, RelativeOffset, Scrollbar};
-use iced::widget::space::horizontal;
+use iced::widget::scrollable::RelativeOffset;
 use iced::widget::{
-    Button, Column, Container, Row, Scrollable, Space, button, canvas, checkbox, column as col,
-    container, pick_list, radio, row, scrollable, text,
+    Button, Column, Row, Scrollable, Space, button, checkbox, column as col, radio, row,
+    scrollable, text,
 };
-use iced::{
-    Alignment, Animation, Border, Color, Element, Font, Length, Pixels, Renderer, Size, Theme,
-    animation, font, window,
-};
+use iced::{Alignment, Border, Color, Element, Length, Pixels, Renderer, Size, Theme, window};
 use iced_aw::style::{Status, menu_bar::primary};
 use iced_aw::widget::InnerBounds;
 use iced_aw::{Menu, MenuBar, iced_aw_font, menu, menu_bar, menu_items, quad};
-use network_interface::Addr::V4;
-use network_interface::{NetworkInterface, NetworkInterfaceConfig};
 use std::sync::LazyLock;
-use std::time::{Duration, Instant};
+use std::time::Instant;
 
 pub struct ConfigWindow {
     pub id: window::Id,
     pub closed: bool,
     pub size: Size,
     pub enable_autoscroll: bool,
-    pub sidebar_state: Animation<bool>,
-    pub bottombar_state: Animation<bool>,
-    pub server_address_text: String,
+    pub sidebar: Sidebar,
+    pub bottombar: BottomBar,
     pub theme: Theme,
-    pub power_button_cache: canvas::Cache,
-    server_address_presets: suggestion_text_input::State<String>,
-}
-
-#[derive(Debug, Clone)]
-pub enum SidebarMessage {
-    Toggle,
-    UpdateAddressPresets,
-}
-
-#[derive(Debug, Clone)]
-pub enum BottomBarMessage {
-    Toggle,
 }
 
 pub static PLAYLIST_SCROLLABLE_ID: LazyLock<iced::widget::Id> =
@@ -64,16 +45,9 @@ impl Window for ConfigWindow {
             size: Size::default(),
 
             enable_autoscroll: true,
-            sidebar_state: Animation::new(false)
-                .duration(Duration::from_millis(100))
-                .easing(animation::Easing::EaseInOut),
-            bottombar_state: Animation::new(false)
-                .duration(Duration::from_millis(100))
-                .easing(animation::Easing::EaseInOut),
-            server_address_presets: suggestion_text_input::State::default(),
-            server_address_text: String::new(),
+            sidebar: Sidebar::new(),
+            bottombar: BottomBar::new(),
             theme: Theme::Dark,
-            power_button_cache: canvas::Cache::default(),
         }
     }
 
@@ -92,20 +66,25 @@ impl Window for ConfigWindow {
 
 impl ConfigWindow {
     pub fn view<'a>(&'a self, dance_interpreter: &'a DanceInterpreter) -> Element<'a, Message> {
-        let top_bar = self.build_top_bar(dance_interpreter);
+        let top_bar = self.build_menu_bar(dance_interpreter);
         let playlist_view = self.build_playlist_view(dance_interpreter);
 
-        let side_bar =
-            self.build_server_sidebar(dance_interpreter)
-                .width(self.sidebar_state.interpolate(
-                    0.0,
-                    (self.size.width / 5.0).min(400.0),
+        let side_bar = self
+            .sidebar
+            .build(dance_interpreter)
+            .width(self.sidebar.state.interpolate(
+                30.0,
+                (self.size.width / 5.0).min(400.0),
+                Instant::now(),
+            ));
+        let bottom_bar =
+            self.bottombar
+                .build(dance_interpreter)
+                .height(self.bottombar.state.interpolate(
+                    80.0,
+                    self.size.height / 3.0,
                     Instant::now(),
                 ));
-        let bottom_bar = self.build_statics_bottombar(dance_interpreter).height(
-            self.bottombar_state
-                .interpolate(60.0, self.size.height / 3.0, Instant::now()),
-        );
 
         col![row![col![top_bar, playlist_view], side_bar], bottom_bar]
             .spacing(5)
@@ -145,112 +124,6 @@ impl ConfigWindow {
         }
 
         (is_current, is_next, is_traktor, is_played)
-    }
-
-    fn build_statics_bottombar<'a>(
-        &'a self,
-        dance_interpreter: &'a DanceInterpreter,
-    ) -> Container<'a, Message> {
-        container(self.build_statics_view(dance_interpreter))
-            .width(Length::Fill)
-            .style(|t: &Theme| {
-                container::Style::default()
-                    .background(t.extended_palette().background.weakest.color)
-            })
-    }
-
-    fn build_server_sidebar<'a>(
-        &'a self,
-        dance_interpreter: &'a DanceInterpreter,
-    ) -> Container<'a, Message> {
-        let sync_options = vec![
-            TraktorSyncMode::None,
-            TraktorSyncMode::Relative,
-            TraktorSyncMode::AbsoluteByNumber,
-            TraktorSyncMode::AbsoluteByName,
-        ];
-
-        let next_options = vec![
-            TraktorNextMode::None,
-            TraktorNextMode::DeckByPosition,
-            TraktorNextMode::DeckByNumber,
-            TraktorNextMode::PlaylistByNumber,
-            TraktorNextMode::PlaylistByName,
-        ];
-
-        container(
-            col![
-                text("Server Settings").size(24),
-                col![
-                    CanvasToggle::new(
-                        dance_interpreter.data_provider.traktor_provider.is_enabled,
-                        &self.power_button_cache
-                    )
-                    .on_toggle(Message::TraktorEnableServer)
-                    .on_draw(power_button::draw_power_button),
-                    text("Enable Server")
-                ],
-                col![
-                    text("Server Address: "),
-                    self.build_network_interface_combo_box(dance_interpreter)
-                ],
-                labeled_message_checkbox(
-                    "Enable Debug Logging",
-                    dance_interpreter
-                        .data_provider
-                        .traktor_provider
-                        .debug_logging,
-                    Message::TraktorEnableDebugLogging,
-                ),
-                label_message_button_fill_opt(
-                    "Reset Connection",
-                    dance_interpreter
-                        .data_provider
-                        .traktor_provider
-                        .is_enabled
-                        .then_some(Message::TraktorReconnect)
-                ),
-                col![
-                    text("Sync Mode"),
-                    pick_list(
-                        sync_options.clone(),
-                        Some(dance_interpreter.data_provider.traktor_provider.sync_mode),
-                        Message::TraktorSetSyncMode
-                    )
-                ]
-                .align_x(Alignment::Center),
-                col![
-                    text("Next Song Mode"),
-                    pick_list(
-                        next_options.clone(),
-                        Some(dance_interpreter.data_provider.traktor_provider.next_mode),
-                        Message::TraktorSetNextMode
-                    )
-                ]
-                .align_x(Alignment::Center),
-                col![
-                    text("Next Song Mode (Fallback)"),
-                    pick_list(
-                        next_options.clone(),
-                        Some(
-                            dance_interpreter
-                                .data_provider
-                                .traktor_provider
-                                .next_mode_fallback
-                        ),
-                        Message::TraktorSetNextModeFallback
-                    )
-                ]
-                .align_x(Alignment::Center)
-            ]
-            .align_x(Alignment::Center)
-            .spacing(10)
-            .padding(10),
-        )
-        .height(Length::Fill)
-        .style(|t| {
-            container::Style::default().background(t.extended_palette().background.weakest.color)
-        })
     }
 
     fn build_playlist_view(&'_ self, dance_interpreter: &DanceInterpreter) -> Column<'_, Message> {
@@ -342,73 +215,6 @@ impl ConfigWindow {
         col!(trow, playlist_scrollable).spacing(5)
     }
 
-    fn build_statics_view<'a>(
-        &self,
-        dance_interpreter: &'a DanceInterpreter,
-    ) -> Scrollable<'a, Message> {
-        let bold_font = Font {
-            family: font::Family::SansSerif,
-            weight: font::Weight::Bold,
-            stretch: font::Stretch::Normal,
-            style: font::Style::Normal,
-        };
-
-        let btn_bottombar_extend = material_icon_sized_message_button(
-            if self.sidebar_state.value() {
-                "bottom_panel_close"
-            } else {
-                "bottom_panel_open"
-            },
-            20.0,
-            Message::Bottombar(BottomBarMessage::Toggle),
-        );
-        let btn_blank: Button<Message> =
-            button(text("Blank").align_y(Vertical::Center).font(bold_font))
-                .style(button::secondary)
-                .on_press(Message::SongChanged(SongChange::Blank));
-        let btn_traktor: Button<Message> =
-            button(text("Traktor").align_y(Vertical::Center).font(bold_font))
-                .style(button::secondary)
-                .on_press(Message::SongChanged(SongChange::Traktor));
-        let mut statics: Vec<Element<_>> = dance_interpreter
-            .data_provider
-            .statics
-            .iter()
-            .enumerate()
-            .map(|(idx, s)| {
-                button(text(&s.dance).font(bold_font))
-                    .style(button::secondary)
-                    .on_press(Message::SongChanged(SongChange::StaticAbsolute(idx)))
-                    .into()
-            })
-            .collect();
-        statics.insert(0, btn_bottombar_extend.into());
-        statics.insert(1, btn_blank.into());
-        statics.insert(2, btn_traktor.into());
-
-        scrollable(row(statics).spacing(5))
-            .direction(Direction::Horizontal(Scrollbar::new()))
-            .spacing(5)
-            .width(Length::Fill)
-    }
-
-    fn build_top_bar<'a>(&self, dance_interpreter: &'a DanceInterpreter) -> Row<'a, Message> {
-        row![
-            self.build_menu_bar(dance_interpreter),
-            horizontal(),
-            material_icon_sized_message_button(
-                if self.sidebar_state.value() {
-                    "right_panel_close"
-                } else {
-                    "right_panel_open"
-                },
-                20.0,
-                Message::Sidebar(SidebarMessage::Toggle)
-            )
-            .padding([0, 4])
-        ]
-    }
-
     fn build_menu_bar<'a>(
         &self,
         dance_interpreter: &'a DanceInterpreter,
@@ -462,76 +268,6 @@ impl ConfigWindow {
 
         mb
     }
-
-    fn build_network_interface_combo_box(
-        &'_ self,
-        dance_interpreter: &DanceInterpreter,
-    ) -> SuggestionTextInput<'_, String, Message> {
-        SuggestionTextInput::new(
-            &self.server_address_presets,
-            if !self.server_address_text.is_empty() {
-                self.server_address_text.as_ref()
-            } else {
-                TRAKTOR_SERVER_DEFAULT_ADDR
-            },
-            Some(&dance_interpreter.data_provider.traktor_provider.address),
-            Message::TraktorChangeAndSubmitAddress,
-        )
-        .on_open(Message::Sidebar(SidebarMessage::UpdateAddressPresets))
-        .on_option_hovered(Message::TraktorChangeAddress)
-        .on_input(Message::TraktorChangeAddress)
-        .on_close(Message::TraktorSubmitAddress)
-    }
-
-    pub fn update_network_interface_selection(&mut self, song_data_provider: &SongDataProvider) {
-        let mut detected_interfaces: Vec<String> =
-            get_formatted_network_interfaces(song_data_provider)
-                .into_iter()
-                .map(|(_, _, formatted)| formatted)
-                .collect();
-        detected_interfaces.push(TRAKTOR_SERVER_DEFAULT_ADDR.to_owned());
-        detected_interfaces.sort();
-
-        self.server_address_presets = suggestion_text_input::State::with_selection(
-            detected_interfaces,
-            Some(&song_data_provider.traktor_provider.address.clone()),
-        );
-    }
-}
-
-fn get_network_interfaces() -> Vec<(String, String)> {
-    let mut interfaces = vec![("any".to_owned(), "0.0.0.0".to_owned())];
-
-    if let Ok(network_interfaces) = NetworkInterface::show() {
-        for i in network_interfaces {
-            for addr in i.addr {
-                let V4(ipv4_addr) = addr else {
-                    continue;
-                };
-
-                interfaces.push((i.name.clone(), ipv4_addr.ip.to_string()));
-            }
-        }
-    }
-
-    interfaces
-}
-
-fn get_formatted_network_interfaces(
-    song_data_provider: &'_ SongDataProvider,
-) -> Vec<(String, String, String)> {
-    let interfaces = get_network_interfaces();
-
-    let original_addr = song_data_provider
-        .traktor_provider
-        .get_socket_addr()
-        .unwrap_or(TRAKTOR_SERVER_DEFAULT_ADDR.parse().unwrap());
-    let original_port = original_addr.port();
-
-    interfaces
-        .into_iter()
-        .map(|(name, addr)| (name, addr.clone(), format!("{}:{}", addr, original_port)))
-        .collect()
 }
 
 fn label_message_button_fill<'a>(
